+++ src/main.ts
@@ -18,10 +18,13 @@ import { HeadlessScreenshotHelperOptions, STYLIZED_POSTPROCESSING, defaultCanvas
 import { setFSModule } from 'molstar/lib/commonjs/mol-util/data-source';
 
 import { PDBeAPI } from './api';
+import { Args, ImageTypes, Mode, Modes } from './args';
 import { collectCaptions } from './captions/collect';
+import { checkMissingFiles, getExpectedFiles } from './expected-files';
 import { fetchUrl, gunzipData, parseIntStrict } from './helpers/helpers';
-import { LogLevel, LogLevels, configureLogging, getLogger, oneLine } from './helpers/logging';
-import { ImageGenerator, ImageType, ImageTypes, Mode, Modes } from './image-generator';
+import { LogLevels, configureLogging, getLogger, oneLine } from './helpers/logging';
+import { ImageGenerator } from './image-generator';
+import * as Paths from './paths';
 import { makeSaveFunction } from './save';
 
 
@@ -37,27 +40,6 @@ const DEFAULT_INPUT_URL_TEMPLATES: { [mode in Mode]: string } = {
 const DEFAULT_PDBE_API_URL = 'https://www.ebi.ac.uk/pdbe/api';
 const DEFAULT_IMAGE_SIZE = '800x800';
 
-
-/** Command line argument values for `main` */
-export interface Args {
-    entry_id: string,
-    output_dir: string,
-    input: string | undefined,
-    input_public: string | undefined,
-    mode: Mode,
-    api_url: string,
-    api_retry: boolean,
-    no_api: boolean,
-    size: { width: number, height: number }[],
-    render_each_size: boolean,
-    type: ImageType[],
-    view: 'front' | 'all' | 'auto',
-    opaque_background: boolean,
-    no_axes: boolean,
-    date: string | undefined,
-    clear: boolean,
-    log: LogLevel,
-}
 
 /** Return parsed command line arguments for `main` */
 export function parseArguments(): Args {
@@ -105,6 +87,7 @@ export async function main(args: Args) {
         }
     }
 
+    let expectedFiles: string[] | undefined = undefined;
     if (args.type.length > 0) {
         const defaultUrl = DEFAULT_INPUT_URL_TEMPLATES[args.mode].replace(/\$\{id\}/g, args.entry_id); // replace ${id} by actual ID
         let runtimeUrl = resolveUrl(args.input) ?? defaultUrl;
@@ -117,11 +100,15 @@ export async function main(args: Args) {
         }
 
         const api = new PDBeAPI(args.api_url, args.no_api, args.api_retry);
+        expectedFiles = await getExpectedFiles(args, api);
+        fs.writeFileSync(Paths.expectedFilelist(args.output_dir, args.entry_id), expectedFiles.join('\n') + '\n');
+        await api.saveCache(Paths.apiDataPath(args.output_dir, args.entry_id));
+
         const plugin = await createHeadlessPlugin(args);
         try {
             const saveFunction = makeSaveFunction(plugin, args.output_dir, args, publicUrl);
             const imageGenerator = new ImageGenerator(plugin, saveFunction, api, args.type, args.view);
-            await imageGenerator.processAll(args.entry_id, runtimeUrl, args.mode, path.join(args.output_dir, `${args.entry_id}_api_data.json`));
+            await imageGenerator.processAll(args.entry_id, runtimeUrl, args.mode);
             if (tmpStructureFile) fs.rmSync(tmpStructureFile, { force: true });
         } finally {
             plugin.dispose();
@@ -131,6 +118,10 @@ export async function main(args: Args) {
     }
 
     collectCaptions(args.output_dir, args.entry_id, args.date);
+    if (expectedFiles) {
+        checkMissingFiles(args.output_dir, expectedFiles, args.entry_id);
+        fs.rmSync(Paths.expectedFilelist(args.output_dir, args.entry_id));
+    }
 }
 
 /** Return a new and initiatized HeadlessPlugin */